+++ lib/danger/commands/init.rb
@@ -255,7 +255,7 @@ module Danger
       ui.say "You need to expose a token called " + "DANGER_GITHUB_API_TOKEN".yellow + " and the value is the GitHub Personal Access Token."
       ui.say "Depending on the CI system, this may need to be done on the machine (in the " + "~/.bashprofile".yellow + ") or in a web UI somewhere."
       ui.say "We have a guide for all supported CI systems on danger.systems:"
-      ui.link "http://danger.systems/guides/getting_started.html#setting-up-danger-to-run-on-your-ci"
+      ui.link "https://danger.systems/guides/getting_started.html#setting-up-danger-to-run-on-your-ci"
     end
 
     def note_about_clicking_links