+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ## master
 
+## 5.6.1
+
 * Use HTTPS links where applicable. [@allewun](https://github.com/allewun)
 * Fix `Danger::GitLabCI` to use `CI_PROJECT_PATH` instead of `CI_PROJECT_ID` for its `#repo_slug`. [@rymai](https://github.com/rymai)